WTE-126 Group Details screen with ListDetails screen. Refactor ListDetails business logic into its own component under functions.
