+++ src/functions/Details/index.js
@@ -1,0 +1,3 @@
+import { handleFilterNutritionCriteria } from './ListDetails';
+
+export { handleFilterNutritionCriteria };
