+++ src/screens/Details/components/index.js
@@ -1,0 +1,3 @@
+import ListDetails from './ListDetails';
+
+export default ListDetails;
