+++ src/screens/index.js
@@ -1,4 +1,3 @@
-import Details from './Details';
 import Feed from './Feed';
 import SavedRecipes from './SavedRecipes';
 import SearchRecipes from './SearchRecipes';
@@ -7,7 +6,6 @@ import ReduxTest from './ReduxTest';
 import SearchResultFeed from './SearchResultFeed';
 
 export {
-  Details,
   Feed,
   SavedRecipes,
   SearchRecipes,
